refactor(query): extract find/find_all engine into query/find
Move node_find/node_find_all and their find-only helpers verbatim out of
dom/element.c into query/find/find.c. Promote the helpers shared with the
select bindings (find_node_attr, append_wrapped, the handle index trio) to
static inline in dom/nodes.h. No behavior change.
